requirements.txt
flask==1.1.1
Flask-RESTful==0.3.8
flask_Cors==3.0.8
Flask-Script==2.0.6
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.3.24
SQLAlchemy-Utils==0.36.8
python-dotenv==0.15.0
python-dateutil==2.8.1
python-editor==1.0.4
pytest==6.2.0
pluggy==0.13.1
py==1.9.0
pytest-cov==2.10.0
pytz==2019.3
pyparsing==2.4.7
pyquery==1.4.1
regex==2020.6.8
geopy==2.0.0
numpy==1.19.2
newspaper3k==0.2.8
gunicorn==20.0.4
requests
pandas
Flask-Mail==0.9.1
Flask-DebugToolbar==0.11.0
Werkzeug==0.16.1
pycountry==20.7.3
debugpy
ptvsd
Uwsgi
beautifulsoup4==4.9.0
click==7.1.1
ez-setup==0.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
scipy==1.4.1
urllib3==1.25.9
validators
Flask-Uploads==0.2.1
flask-marshmallow==0.11.0
marshmallow==3.5.1
marshmallow-sqlalchemy==0.22.3
psycopg2==2.8.4
Faker==4.1.1
flake8==3.8.3
wget==3.2
selectorlib==0.16.0
multiprocess==0.70.11.1
joblib==0.14.1
tenacity==7.0.0
copyheaders==0.0.2
glom==20.11.0
fake-useragent
word2number==1.1
googletrans==3.1.0a0
pycountry==20.7.3
pycountry-convert==0.7.2
psutil==5.8.0
modin[all]==0.11.0